+++ src/alice.ts
@@ -18,6 +18,9 @@ import {
 } from './middlewares'
 
 import aliceStateMiddleware from './middlewares/aliceStateMiddleware'
+import { configInterface } from './types/alice'
+import { CommandInterface } from './types/command'
+import { WebhookResponse, WebhookRequest } from 'webhook'
 
 const DEFAULT_SESSIONS_LIMIT: number = 1000
 
@@ -29,10 +32,12 @@ export default class Alice {
   private scenes: Scene[]
   private currentScene: Scene | null
   private sessions: Sessions
-  private config: {}
-  private server: {}
+  private server: {
+    close: () => void,
+  }
+  private config: configInterface
 
-  constructor(config = {}) {
+  constructor(config: configInterface = {}) {
     this.anyCallback = null
     this.welcomeCallback = null
     this.commands = new Commands(config.fuseOptions || null)
@@ -107,6 +112,23 @@ export default class Alice {
     const sessionId = selectSessionId(req)
     const session = this.sessions.findOrCreate(sessionId)
 
+    /*
+     * Initializing context of the request
+     */
+    const ctxDefaultParams = {
+      req,
+      session,
+      sendResponse: sendResponse || null,
+      /*
+       * if Alice is listening on express.js port, add this server instance
+       * to the context
+       */
+      server: this.server || null,
+      middlewares: this.middlewares,
+    }
+    const ctxInstance = new Ctx(ctxDefaultParams)
+    const ctxWithMiddlewares = await applyMiddlewares(this.middlewares, ctxInstance)
+
     /* check whether current scene is not defined */
     if (!session.getData('currentScene')) {
       session.setData('currentScene', null)
@@ -124,12 +146,12 @@ export default class Alice {
        */
       if (matchedScene) {
         if (matchedScene.isLeaveCommand(requestedCommandName)) {
-          matchedScene.handleRequest(req, sendResponse, session)
+          await matchedScene.handleRequest(req, sendResponse, ctxWithMiddlewares)
           session.setData('currentScene', null)
           return true
         } else {
           const sceneResponse = await matchedScene.handleRequest(
-            req, sendResponse, session,
+            req, sendResponse, ctxWithMiddlewares,
           )
           if (sceneResponse) {
             return true
@@ -145,7 +167,7 @@ export default class Alice {
       if (matchedScene) {
         session.setData('currentScene', matchedScene.name)
         const sceneResponse = await matchedScene.handleRequest(
-          req, sendResponse, session,
+          req, sendResponse, ctxWithMiddlewares,
         )
         if (sceneResponse) {
           return true
@@ -153,23 +175,7 @@ export default class Alice {
       }
     }
 
-    const requestedCommands = this.commands.search(requestedCommandName)
-
-    /*
-     * Initializing context of the request
-     */
-    const ctxDefaultParams = {
-      req,
-      session,
-      sendResponse: sendResponse || null,
-      /*
-       * if Alice is listening on express.js port, add this server instance
-       * to the context
-       */
-      server: this.server || null,
-      middlewares: this.middlewares,
-    }
-
+    const requestedCommands = await this.commands.search(ctxWithMiddlewares)
     /*
     * Если новая сессия, то запускаем стартовую команду
     */
@@ -178,10 +184,6 @@ export default class Alice {
        * Patch context with middlewares
        */
       if (this.welcomeCallback) {
-        // tslint:disable:no-shadowed-variable
-        const ctxInstance = new Ctx(ctxDefaultParams)
-        const ctxWithMiddlewares = await applyMiddlewares(this.middlewares, ctxInstance)
-        // tslint:enable:no-shadowed-variable
         return await this.welcomeCallback(ctxWithMiddlewares)
       }
     }
@@ -190,13 +192,8 @@ export default class Alice {
      * Запускаем её обработчик.
      */
     if (requestedCommands.length !== 0) {
-      const requestedCommand = requestedCommands[0]
-      // tslint:disable:no-shadowed-variable
-      const ctxInstance = new Ctx(merge(ctxDefaultParams, {
-        command: requestedCommand,
-      }))
-      const ctxWithMiddlewares = await applyMiddlewares(this.middlewares, ctxInstance)
-      // tslint:enable:no-shadowed-variable
+      const requestedCommand: CommandInterface = requestedCommands[0]
+      ctxWithMiddlewares.command = requestedCommand
       return await requestedCommand.callback(ctxWithMiddlewares)
     }
 
@@ -204,9 +201,6 @@ export default class Alice {
      * Такой команды не было зарегестрировано.
      * Переходим в обработчик исключений
      */
-    const ctxInstance = new Ctx(ctxDefaultParams)
-    const ctxWithMiddlewares = await applyMiddlewares(this.middlewares, ctxInstance)
-
     if (!this.anyCallback) {
       throw new Error([
         `alice.any(ctx => ctx.reply('404')) Method must be defined`,
@@ -219,7 +213,10 @@ export default class Alice {
   /*
    * Same as handleRequestBody, but syntax shorter
    */
-  public async handleRequest(req, sendResponse) {
+  public async handleRequest(
+    req: WebhookRequest,
+    sendResponse?: (res: WebhookResponse) => void,
+  ): Promise<any> {
     return await this.handleRequestBody(req, sendResponse)
   }
 
@@ -231,7 +228,7 @@ export default class Alice {
    * При получении ответа от @handleRequestBody, результат
    * отправляется обратно.
    */
-  public async listen(callbackUrl = '/', port = 80, callback: () => void) {
+  public async listen(callbackUrl = '/', port = 80, callback?: () => void) {
     return new Promise((resolve) => {
       const app = express()
       app.use(express.json())
@@ -266,10 +263,10 @@ export default class Alice {
     }
   }
 
-  private _handleEnterScene(sceneName) {
+  protected _handleEnterScene(sceneName) {
     this.currentScene = sceneName
   }
-  private _handleLeaveScene(sceneName) {
+  protected _handleLeaveScene() {
     this.currentScene = null
   }
 }