+++ src/types/alice.d.ts
@@ -1,0 +1,4 @@
+export interface configInterface {
+  fuseOptions?: {},
+  sessionsLimit?: number
+}