+++ src/types/button.d.ts
@@ -1,0 +1,16 @@
+export interface ButtonParams {
+  title?: string,
+  text?: string,
+  tts?: string,
+  url?: string,
+  hide?: boolean,
+  payload?: {}
+}
+
+export interface ButtonInterface {
+  title: string,
+  tts?: string,
+  url?: string,
+  hide?: boolean,
+  payload?: {}
+}