+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+<a name="1.3.3"></a>
+## [1.3.3](https://github.com/fletcherist/yandex-dialogs-sdk/compare/v1.3.2...v1.3.3) (2018-07-08)
+
+
+### Bug Fixes
+
+* bug with commands (array) ([8cac580](https://github.com/fletcherist/yandex-dialogs-sdk/commit/8cac580))
+* package.json ([169a818](https://github.com/fletcherist/yandex-dialogs-sdk/commit/169a818))
+
+
+### Features
+
+* add guess number game to examples ([3d34c26](https://github.com/fletcherist/yandex-dialogs-sdk/commit/3d34c26))
+* add youtube link to readme ([31c8328](https://github.com/fletcherist/yandex-dialogs-sdk/commit/31c8328))
+
+
+
+<a name="1.3.2"></a>
+## [1.3.2](https://github.com/fletcherist/yandex-dialogs-sdk/compare/v1.2.1...v1.3.2) (2018-07-08)
+
+
+### Bug Fixes
+
+* bug with generate request func ([570b23c](https://github.com/fletcherist/yandex-dialogs-sdk/commit/570b23c))
+* button.d.ts error ([44ff44b](https://github.com/fletcherist/yandex-dialogs-sdk/commit/44ff44b))
+* ctx types ([ff44115](https://github.com/fletcherist/yandex-dialogs-sdk/commit/ff44115))
+* fixed replyBuilder types ([72d2814](https://github.com/fletcherist/yandex-dialogs-sdk/commit/72d2814))
+* invalid import replyBuilder ([c98f048](https://github.com/fletcherist/yandex-dialogs-sdk/commit/c98f048))
+* reconfigure jest ([31f5b72](https://github.com/fletcherist/yandex-dialogs-sdk/commit/31f5b72))
+* session types ([a646c47](https://github.com/fletcherist/yandex-dialogs-sdk/commit/a646c47))
+* types in alice.ts ([2a4e266](https://github.com/fletcherist/yandex-dialogs-sdk/commit/2a4e266))
+* typos in tests, ctx and main ([0c55743](https://github.com/fletcherist/yandex-dialogs-sdk/commit/0c55743))
+
+
+### Features
+
+* add types, implement confirm middleware draft ([527c715](https://github.com/fletcherist/yandex-dialogs-sdk/commit/527c715))
+* add webhook request/response types ([cee8ec7](https://github.com/fletcherist/yandex-dialogs-sdk/commit/cee8ec7))
+* added another project into readme ([09c04f8](https://github.com/fletcherist/yandex-dialogs-sdk/commit/09c04f8))
+* fix default callback null ([21a0117](https://github.com/fletcherist/yandex-dialogs-sdk/commit/21a0117))
+* implement aliceStateMiddleware ([ca4cd35](https://github.com/fletcherist/yandex-dialogs-sdk/commit/ca4cd35))
+* implement middlewares logic ([4b3c1b1](https://github.com/fletcherist/yandex-dialogs-sdk/commit/4b3c1b1))
+* implement modules' interfaces ([c285250](https://github.com/fletcherist/yandex-dialogs-sdk/commit/c285250))
+* implement typescript support, change file extension names to ts, fix few types ([6dd23b1](https://github.com/fletcherist/yandex-dialogs-sdk/commit/6dd23b1))
+* improve alice.ts ([4afa771](https://github.com/fletcherist/yandex-dialogs-sdk/commit/4afa771))
+* improve applyMiddlewares mechanism ([6991279](https://github.com/fletcherist/yandex-dialogs-sdk/commit/6991279))
+* update readme ([021ecbe](https://github.com/fletcherist/yandex-dialogs-sdk/commit/021ecbe))
+* update readme ([3692de1](https://github.com/fletcherist/yandex-dialogs-sdk/commit/3692de1))
+
+
+
 <a name="1.2.1"></a>
 ## [1.2.1](https://github.com/fletcherist/yandex-dialogs-sdk/compare/v1.2.0...v1.2.1) (2018-07-01)